+++ src/main/java/org/matsim/prepare/DownloadUseCase.java
@@ -1,0 +1,4 @@
+package org.matsim.prepare;
+
+public class DownloadUseCase {
+}
