+++ src/main/java/org/matsim/prepare/ManualSampling.java
@@ -1,0 +1,4 @@
+package org.matsim.prepare;
+
+public class ManualSampling {
+}
