+++ src/main/java/org/matsim/prepare/RandomSampling.java
@@ -1,0 +1,4 @@
+package org.matsim.prepare;
+
+public class RandomSampling {
+}
