+++ src/main/java/org/matsim/utils/ScenariosTools.java
@@ -1,0 +1,8 @@
+package org.matsim.utils;
+
+public class ScenariosTools {
+    public enum RawScenarios {BERLIN, LEIPZIG, MANHATTAN_TAXI, VULKANEIFEL, KELHEIM}
+
+    public enum UseCases {BERLIN_DRT, LEIPZIG_DRT, VULKANEIFEL_SCHOOL_TRANSPORT, KELHEIM_KEXI, ORANIENBURG}
+
+}
